test: game - ⭕ InputParser
