+++ project/app/src/test/java/com/orange/labs/orangetrainingbox/MockUtils.kt
@@ -18,6 +18,7 @@
 package com.orange.labs.orangetrainingbox
 
 import android.bluetooth.BluetoothDevice
+import android.bluetooth.BluetoothGattCharacteristic
 import android.view.View
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito
@@ -69,6 +70,46 @@ class MockUtils {
             `when`(view.width).thenReturn(width)
             `when`(view.height).thenReturn(height)
             return view
+        }
+
+        /**
+         * Creates using Mockito library a mock object for class BluetoothGattCharacteristic.
+         *
+         * This mock will return the same frame with raw values.
+         * The format of the frame follows implementation of the v2.0.0 protocol of Arduino firmware embedded in
+         * the Baah Box (https://github.com/Orange-OpenSource/BaahBox-Arduino/releases).
+         *
+         * Thus the frame contains 6 bytes like:
+        <pre>
+        It models data like <muscle1, muscle2, Joystic=JBin, EndOfFrame>
+        Where:
+        - muscle1 = C1 x 32 + a1
+        - muscle2 = C2 x 32 + a2
+        - joystic = JBin
+        - EndOfFrame = 90 -> '\n'
+        </pre>
+         *
+         * See: https://site.mockito.org/
+         *
+         * @param c1 Used to compute the muscle 1 value, multiplied by 32
+         * @param a1 Added to result of c1 x 32 to compute muscle 1 value
+         * @param c2 Used to compute the muscle 2 value, multiplied by 32
+         * @param a2 Added to result of c2 x 32 to compute muscle 2 value
+         * @param joystick Joystick value
+         * @return BluetoothGattCharacteristic The mocked object
+         */
+        fun mockBluetoothGattCharacteristic(c1: Int, a1: Int, c2: Int, a2: Int, joystick: Int): BluetoothGattCharacteristic {
+
+            val mock: BluetoothGattCharacteristic = mock(BluetoothGattCharacteristic::class.java)
+
+            `when`(mock.getIntValue(BluetoothGattCharacteristic.FORMAT_UINT8,0)).thenReturn(c1)
+            `when`(mock.getIntValue(BluetoothGattCharacteristic.FORMAT_UINT8,1)).thenReturn(a1)
+            `when`(mock.getIntValue(BluetoothGattCharacteristic.FORMAT_UINT8,2)).thenReturn(c2)
+            `when`(mock.getIntValue(BluetoothGattCharacteristic.FORMAT_UINT8,3)).thenReturn(a2)
+            `when`(mock.getIntValue(BluetoothGattCharacteristic.FORMAT_UINT8,4)).thenReturn(joystick)
+
+            return mock
+
         }
 
     }
